ab-sql-final-project/telecom_project/dashboard/requirements.txt
pandas
dash
dash_core_components
dash_html_components
dash.dependencies
plotly.graph_objs